% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini.R
\name{gini}
\alias{gini}
\title{Gini index}
\usage{
gini(dataset, ipuc = "ipuc", hhcsw = "DB090", hhsize = "HX040",
  ci = NULL, rep = 1000, verbose = FALSE)
}
\arguments{
\item{dataset}{a data.frame containing variables obtained by using the setupDataset function.}

\item{ipuc}{a character string indicating the variable name of the income per unit of consumption within dataset. Default is "ipuc".}

\item{hhcsw}{a character string indicating the variable name of the household cross-sectional weight within dataset. Default is "DB090".}

\item{hhsize}{a character string indicating the variable name of the household size within dataset. Default is "HX040".}

\item{ci}{a scalar or vector containing the confidence level(s) of the required interval(s). Default does not calculate the confidence interval.}

\item{rep}{a number to do the confidence interval using boostrap technique.}

\item{verbose}{logical; if TRUE the confindence interval is plotted.}
}
\value{
The value of the Gini index.
}
\description{
Estimates the Gini inequality index of an income distribution.
}
\details{
The Gini index is calculated using the equivalized disposable income of each individual. Two types of equivalence scales can be used, the modified OECD scale and the parametric scale of Buhmann et al. (1988). The default is the modified OECD scale  (see setupDataset).
}
\examples{
data(eusilc2)
ATdataset <- setupDataset(eusilc2, country = "AT", s = "OECD")
gini(ATdataset)

}
\references{
B. Buhmann et al. (1988) Equivalence scales, well-being, inequality and poverty: sensitivity estimates across ten countries using the Luxembourg Income Study (LIS) database, Review of Income and Wealth, 34, 115--142.

E. Ferreira and A. Garín (1997) Una nota sobre el cálculo del índice de Gini, Estadística Española, 39(142), 207--218.

\url{http://ec.europa.eu/eurostat/statistics-explained/index.php/Glossary:Gini_coefficient}
}
\seealso{
setupDataset
}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo
}
