% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterEUSILC.R
\name{filterEUSILC}
\alias{filterEUSILC}
\alias{filtereusilc}
\title{ToDo}
\usage{
filterEUSILC(year = "2006", varD = c("DB010", "DB020", "DB030", "DB040",
  "DB090"), varH = c("HB010", "HB020", "HB030", "HY020", "HY025", "HX010",
  "HX040", "HX050", "HX080", "HX090"))
}
\arguments{
\item{year}{is the year when the survey was done.}

\item{varD}{are a vector containing the DB variables to filter}
}
\value{
a data frame containing the variables required
}
\description{
This is an auxiliary function to obtain specific variables from EUSILC datasets files.
}
\details{
Filter datasets from EUSILC survey

\code{filterEUSILC} uses some packages
to do the

Todo
}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo
}

