% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCS2012-data.R
\docType{data}
\name{LCS2012}
\alias{LCS2012}
\title{Living conditions survey
CAMBIAR}
\format{A data frame with 6000 rows of 7 variables:
\describe{
 \item{DB010}{integer; year of the pull}
 \item{DB020}{factor; name of the country}
 \item{DB040}{factor; name of the region within the country}}}
\usage{
data(LCS2012)
}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{http://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{datasets}

