% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mih.R
\name{mih}
\alias{mih}
\title{mih}
\usage{
mih(dataset, ci = FALSE, rep = 1000, verbose = FALSE)
}
\description{
This is mean income household function
}
\details{
Todo
}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo
}

