% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mip.R
\name{mip}
\alias{mip}
\title{mip}
\usage{
mip(dataset, ci = FALSE, rep = 1000, verbose = FALSE)
}
\description{
This is mean income person function
}
\details{
Todo
}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo
}

