% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmegaTIP.R
\name{OmegaTIP}
\alias{OmegaTIP}
\title{Matrix for testing TIP dominance}
\usage{
OmegaTIP(dataset, arpt.value, norm = FALSE, samp)
}
\arguments{
\item{dataset}{a data.frame containing variables obtained by using the setupDataset function.}

\item{arpt.value}{the at-risk-of-poverty threshold to be used (see arpt).}

\item{norm}{logical; if  TRUE, the normalized TIP curve ordinates are computed using the normalized poverty gaps (poverty gaps divided by the poverty threshold).}

\item{samp}{an integer which represents the number of TIP curve ordinates to be estimated. These ordinates will be estimated at points p_i, where p_i=i/samp, i=1, ..., samp.}
}
\value{
A list with the following components:

Omega covariance matrix for the estimated vector of TIP curve ordinates.

tip.curve estimated vector of TIP curve ordinates.
}
\description{
The auxiliary function OmegaTIP computes the (empirical) vector of TIP curve ordinates and its corresponding covariance matrix. This matrix will be used to compute the test-statistic to test for the TIP dominance relation between two TIP curves.
}
\details{
Estimation of TIP curve ordinates and their covariance matrix are made following Beach and Davidson (1983) and Xu and Osberg (1998).

Calculations are made using the equivalized disposable income. The equivalence scales that can be employed are the modified OECD scale or the parametric scale of Buhmann et al. (1988). The default is the modified OECD scale (see setupDataset).
}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo
}
\references{
C. M. Beach and R. Davidson (1983) Distribution-free statistical inference with Lorenz curves and income shares, Review of Economic Studies, 50, 723--735.

B. Buhmann et al. (1988) Equivalence scales, well-being, inequality and poverty: sensitivity estimates across ten countries using the Luxembourg Income Study (LIS) database, Review of Income and Wealth, 34, 115--142.

K. Xu and L. Osberg (1998) A distribution-free test for deprivation dominance, Econometric Reviews,17, 415--429.
}
\seealso{
testTIP, setupDataset, arpt
}

